+++ src/Libraries/Abstractions/Nultien.TheShop.IServices/IShopService.cs
@@ -19,7 +19,7 @@ namespace Nultien.TheShop.Services
         /// <param name="quantity"></param>
         /// <param name="maxExpectedPrice"></param>
         /// <returns></returns>
-        List<OrderItem> SellArticle(string articleCode, long quantity, float maxExpectedPrice);
+        List<OrderItem> SellArticle(string articleCode, long quantity, double maxExpectedPrice);
 
         /// <summary>
         /// Create order based on order items for customer id.
